requirements.txt
numpy>=1.24
scipy>=1.10
scikit-learn>=1.3
umap-learn>=0.5
hdbscan>=0.8
pandas>=2.0
